# [Mono.Android] fix `Type.GetType()` performance issue