--- a/src/Mono.Android/java/mono/android/TypeManager.java
+++ b/src/Mono.Android/java/mono/android/TypeManager.java
@@ -14,7 +14,7 @@ public class TypeManager {
 		String methods = 
 			"n_activate:(Ljava/lang/String;Ljava/lang/String;Ljava/lang/Object;[Ljava/lang/Object;)V:GetActivateHandler\n" +
 			"";
-		mono.android.Runtime.register ("Java.Interop.TypeManager+JavaTypeManager, Mono.Android, Version=0.0.0.0, Culture=neutral, PublicKeyToken=null", TypeManager.class, methods);
+		mono.android.Runtime.register ("Java.Interop.TypeManager+JavaTypeManager, Mono.Android", TypeManager.class, methods);
 	}
 //#MARSHAL_METHODS:END - do not remove or modify this line, it is required during application build
 }